Added object caching (Fixes Krakensbane lag)
Normally when an object is out of range the game will disable the renderer
object to prevent uneeded strain on the GPU, however the object still
exists in the world and is still being processed. This change disables the
object completely when it is out of the visibility range so that it isn't
processed by the game when it isn't needed.